% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_functs.R
\name{addpoly_acs}
\alias{addpoly_acs}
\title{Add polygons and legend for ACS variable}
\usage{
addpoly_acs(
  leafletobj,
  colvisualize,
  labtext,
  acsdata = leaflet::getMapData(leafletobj),
  totcol = estimate_Population,
  colors = "Blues",
  title = NULL
)
}
\arguments{
\item{leafletobj}{Leaflet obj is base leaflet map adding layer to}

\item{colvisualize}{String name of column visualizing}

\item{labtext}{Text to supply to leafletwrappers::label_output() function to generate labels for dataset}

\item{acsdata}{spatial data visualizing in this polygon. Should be a spatial dataset with columns for data of interest. Default dataset fed to map object}

\item{totcol}{total column/universe of column visualizing (e.g., tot_population, tot_households)}

\item{colors}{Color scheme to shade polygons by}

\item{title}{Title of layer/legend}
}
\value{
Leaflet object with polygons shaded/added
}
\description{
Standard function for visualizng acs variables in web-maps
}
