% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_functs.R
\name{addpoly_race}
\alias{addpoly_race}
\title{Add polygons for ACS5 race data}
\usage{
addpoly_race(
  leafletobj,
  colvisualize,
  acsdata = leaflet::getMapData(leafletobj)
)
}
\arguments{
\item{leafletobj}{leaflet object adding polygons to}

\item{colvisualize}{Column visualizing in polygons (e.g., Hispanic_estimate)}

\item{acsdata}{spatial dataset column is in; default to pull data from map object}
}
\value{
Leaflet map with polygons added
}
\description{
Adds polygons for ACS5 race data to leaflet map.
}
