% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{downloadgeogs}
\alias{downloadgeogs}
\title{Download shapefiles associated with geographies}
\usage{
downloadgeogs(
  geogs = c("block group", "tract", "place", "county"),
  tidyyear = 2020,
  tidystate = "Maryland",
  dirsave = "./data"
)
}
\arguments{
\item{geogs}{Set of geographies to read in and save data for; default c("block group", "tract", "place", "county")}

\item{tidyyear}{Year of geography data to download; default 2020 (for 2016-2020 data)}

\item{tidystate}{State to download data for; default "Maryland"}

\item{dirsave}{Directory to save geometries in}
}
\value{
Saved shapefiles for Census geographies
}
\description{
Download shapefiles associated with geographies
}
