% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_functs.R
\name{gen_acsmap_overall}
\alias{gen_acsmap_overall}
\title{Generate map of key ACS variables at a given geography}
\usage{
gen_acsmap_overall(baseacsdata, pct_est = "pct_", inccountycontrols = T, ...)
}
\arguments{
\item{baseacsdata}{Spatial dataframe for a Census geometry pre-processed by acsmapping::prepallinone_acsmapoverall() sequence of functions}

\item{pct_est}{Percent or estimate prefix as string; default "pct_"}

\item{inccountycontrols}{Whether to include overlay layers for county and place boundaries; default TRUE}

\item{...}{Additional arguments to leafletwrappers::layercontrolsquick.}
}
\value{
A leaflet map with key ACS variables visualized
}
\description{
Generates leaflet map of key ACS variables--including population by race/ethnicity, poverty status, and renter occupancy--for processed ACS dataframe.
}
