% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{loadtidydata_acsmapoverall}
\alias{loadtidydata_acsmapoverall}
\title{Load and Save tidycensus Data for ACS Map Overall}
\usage{
loadtidydata_acsmapoverall(
  geogs = c("block group", "tract", "place", "county"),
  tidyyear = 2020,
  tidystate = "Maryland",
  downloadgeog = F,
  basedir = "./data",
  .geomyear = tidyyear
)
}
\arguments{
\item{geogs}{Set of geographies to read in and save data for; default c("block group", "tract", "place", "county")}

\item{tidyyear}{Last year of 5-year ACS data; default 2020 (for 2016-2020 data)}

\item{tidystate}{State to download data for; default "Maryland"}

\item{downloadgeog}{Whether to download and save spatial geographies associated with data downloading; default F}

\item{basedir}{The base-directory to save data and geometry files in; default "./data/". Creates directory if does not exist}

\item{.geomyear}{If downloadgeog is TRUE, the year of the geometry file to download; default tidyyear}
}
\value{
Reads in and saves tidycensus downloaded ACS data at different geographies to standard file structure
}
\description{
Reads in and saves 5-year tidycensus data to be visualized by acsmapping::gen_acsmap_overall, saving them in a common directory structure for processing by acsmapping::readprocessdata_acsmapoverall() or acsmapping::processdfs_acsmapoverall().
}
\details{
Sequence of functions to prepare data for acsmapping::gen_acsmap_overall() is 1) acsmapping::loadtidydata_acsmapoverall() 2) acsmapping::readprocessdata_acsmapoverall() or acsmapping::processdfs_acsmapoverall() 3) acsmapping::readsave_spatial_acsmapoverall().
}
