% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{pctna_recode}
\alias{pctna_recode}
\title{Recode percent-column to be NA if less than 10 total observations}
\usage{
pctna_recode(df, totcol, colname)
}
\arguments{
\item{df}{ACS5 processed dataset}

\item{totcol}{Column with denominator of percentage column (total), as string}

\item{colname}{Name of percentage column, as string}
}
\value{
Dataframe with percentage column set to NA if less than 10 total observations
}
\description{
Recodes a percentage column to be NA if there are less than 10 observations in the denominator of the percentage column (the total column of the dataset)
}
