% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{pivotfunct}
\alias{pivotfunct}
\title{Pivot-wider acs-data}
\usage{
pivotfunct(df, namescol)
}
\arguments{
\item{df}{Tidycensus-processed dataframe pivoting to join to spatial file. Data should have been processed with corresponding acsprocess package function for dataframe}

\item{namescol}{Column containing groups that are being pivoted wider (e.g., "agegrp", "incpov_ratio"), as string}
}
\value{
Pivoted dataframe with each group as column
}
\description{
Function to pivot acs5 processed-data for joining to spatial file. Leaflet works better with wide than short-data, so pivot data so that totals are associated with each group in the data
}
