% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{processdfs_acsmapoverall}
\alias{processdfs_acsmapoverall}
\title{Process dataframes that build ACS Map Overall}
\usage{
processdfs_acsmapoverall(
  geog,
  tidyyear = 2020,
  tidystate = "Maryland",
  basedir = "./data"
)
}
\arguments{
\item{geog}{Tidycensus geography, as a string (e.g., "block group", "place", "county")}

\item{tidyyear}{Last year of 5-year ACS data; default 2020 (for 2016-2020 data)}

\item{tidystate}{State data assocaited with; default "Maryland"}

\item{basedir}{Base-directory geometry located in; default "./data"}
}
\value{
Saves set of processed-tidycensus dataframes at a given geography as RDS files in standard directory structure
}
\description{
Reads in, processes, and saves unprocessed tidycensus dataframes that make up the ACS Map Overall. Specifically, processes set of dataframes saved by acsmapping::loadtidydata_acsmapoverall() at the same geography in sequence of steps enabling datasets to be visualized by acsmapping::gen_acsmap_overall().
}
\details{
Sequence of functions to prepare data for acsmapping::gen_acsmap_overall() is 1) acsmapping::loadtidydata_acsmapoverall() 2) acsmapping::readprocessdata_acsmapoverall() or acsmapping::processdfs_acsmapoverall() 3) acsmapping::readsave_spatial_acsmapoverall().
}
