% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{quickdircreate}
\alias{quickdircreate}
\title{Check if directory exists and create it if not}
\usage{
quickdircreate(dirpath)
}
\arguments{
\item{dirpath}{Path to directory}
}
\description{
Check if directory exists and create one if not.
}
