% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_functs.R
\name{quicklab}
\alias{quicklab}
\title{Quick label for ACS5 data}
\usage{
quicklab(colsuffix, tot = "estimate_Population")
}
\arguments{
\item{colsuffix}{String-suffix of column visualizing in data. E.g., estimate_Under.poverty.line. Assumes common format of estimate and percentage columns with estimate_{colsuffix} and pct_{colsuffix}}

\item{tot}{Name of total column for universe of variable as string (e.g., if total renter-households, column for total households). Default "estimate_Population"}
}
\value{
Text to be supplied to a leafletwrappers label function for a spatial ACS5 dataframe
}
\description{
Function to generate label text for ACS5 spatial data
}
