% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{readprocessdata_acsmapoverall}
\alias{readprocessdata_acsmapoverall}
\title{Read and Process Tidycensus Data for Multiple Geographies}
\usage{
readprocessdata_acsmapoverall(
  geogs = c("block group", "tract", "place", "county"),
  tidyyear = 2020,
  tidystate = "Maryland",
  basedir = "./data"
)
}
\arguments{
\item{geogs}{Tidycensus geographies, as a string vector. Default c("block group", "tract", "place", "county")}

\item{tidyyear}{Last year of 5-year ACS data; default 2020 (for 2016-2020 data)}

\item{tidystate}{State to download data for; default "Maryland"}

\item{basedir}{The base-directory to data and geometry files are in; default "./data/"}
}
\value{
Saves set of processed-tidycensus dataframes at a given set of geographies as RDS files in standard directory structure
}
\description{
Reads in, processes, and saves unprocessed tidycensus dataframes that make up the ACS Map Overall, for a given set of geographies. Specifically, processes set of dataframes saved by acsmapping::loadtidydata_acsmapoverall() at the same geography in sequence of steps enabling datasets to be visualized by acsmapping::gen_acsmap_overall().
}
