% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{readsave_spatial_acsmapoverall}
\alias{readsave_spatial_acsmapoverall}
\title{Generate shapefile for ACS map overall function}
\usage{
readsave_spatial_acsmapoverall(
  geog,
  tidyyear = 2020,
  tidystate = "Maryland",
  basedir = "./data"
)
}
\arguments{
\item{geog}{Tidycensus geography, as a string (e.g., "block group", "tract", "place", "county")}

\item{tidyyear}{Last year of 5-year ACS data; default 2020 (for 2016-2020 data)}

\item{tidystate}{State data is for; default "Maryland"}

\item{basedir}{Base directory state geography file is in; default "./data"}
}
\value{
Saves shapefile with tidycensus data joined-to-it ready for visualization by acsmapping::gen_acsmap_overall()
}
\description{
Creates shapefile for operation with acsmapping::gen_acsmap_overall() function, joining procssed ACS datasets at the same geometry saved by acsmapping::processdfs_acsmapoverall() or readprocessdata_acsmapoverall() to an ACS shapefile for visualization by gen_acsmap_overall().
}
